Began Strategy Pattern
I created the StyleStrategy interface that requires a method to draw the actual game board with a given color for the board, color for the pit, and shape. We still need to work on exactly how the Shape will be selected. Waiting on an email from the professor to continue.